+++ plugins/tests/aws/test_hpo.py
@@ -12,7 +12,7 @@ from flytekit.models.sagemaker.hpo_job import (
 )
 from flytekit.models.sagemaker.parameter_ranges import IntegerParameterRange, ParameterRangeOneOf
 from flytekit.models.sagemaker.training_job import AlgorithmName, AlgorithmSpecification, TrainingJobResourceConfig
-from plugins.aws.tests.test_training import _get_reg_settings
+from plugins.tests.aws.test_training import _get_reg_settings
 
 
 def test_hpo_for_builtin():
