+++ plugins/tests/aws/test_training.py
